feat: setAttribute 해주는 메서드에서 create 까지 해주게 업그레이드
